+++ ml_tools/optimization_tools.py
@@ -2,6 +2,7 @@ from ._core._optimization_tools import (
     create_optimization_bounds,
     parse_lower_upper_bounds,
     plot_optimal_feature_distributions,
+    plot_optimal_feature_distributions_from_dataframe,
     info
 )
 
@@ -9,4 +10,5 @@ __all__ = [
     "create_optimization_bounds",
     "parse_lower_upper_bounds",
     "plot_optimal_feature_distributions",
+    "plot_optimal_feature_distributions_from_dataframe",
 ]